addressed comments

+++ src/libbson/src/bson/bson-keys.c
@@ -143,9 +143,8 @@ bson_uint32_to_string (uint32_t value,      /* IN */
    *strptr = str;
 
    int ret = bson_snprintf (str, size, "%u", value);
-   // Truncation is OK. Assert no error occurred.
+   // Truncation is OK. 
    BSON_ASSERT (ret > 0);
-   // Check in bounds of a size_t.
    BSON_ASSERT (bson_in_range_size_t_signed (ret));
-   return (int) ret;
+   return (size_t) ret;
 }
